requirements.txt
# install torch & torchvision versions that works
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
Pillow>=8.0.0
tqdm>=4.60.0
datasets

# For graph neural networks
torch-geometric>=2.0.0
pytorch-lightning>=1.5.0

# For topological data analysis
gudhi>=3.5.0
ripser>=0.6.0

# For image processing
opencv-python>=4.5.0
scikit-image>=0.18.0

# For analysis and evaluation
scikit-learn>=1.0.0
seaborn>=0.11.0

# For Jupyter notebooks
ipywidgets>=7.6.0
jupyter>=1.0.0
ipython>=7.0.0

# Utilities
requests>=2.25.0
kagglehub